# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SSC Vocabulary Master API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tests:
    post:
      operationId: createTest
      tags: [tests]
      summary: Create a new test
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestConfig"
      responses:
        "201":
          description: Test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestWithQuestions"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tests/history:
    get:
      operationId: getTestHistory
      tags: [tests]
      summary: Get user test history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Test history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestHistoryPage"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tests/{id}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get test with questions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestWithQuestions"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tests/{id}/submit:
    post:
      operationId: submitTest
      tags: [tests]
      summary: Submit test answers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestSubmission"
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tests/{id}/result:
    get:
      operationId: getTestResult
      tags: [tests]
      summary: Get test result
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Progress ────────────────────────────────────────────────────────────
...
  /stats/dashboard:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get user dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stats/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [stats]
      summary: Get top performers leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    VocabularyItem:
      type: object
      required: [id, word, meaning, category, difficulty, alphabet]
      properties:
        id:
          type: integer
        word:
          type: string
        meaning:
          type: string
        hindiMeaning:
          type: ["string", "null"]
        exampleSentence:
          type: ["string", "null"]
        synonyms:
          type: array
          items:
            type: string
        antonyms:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        category:
          type: string
        alphabet:
          type: string
        topics:
          type: array
          items:
            type: string
        examRefs:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    TestConfig:
      type: object
      required: [totalQuestions]
      properties:
        categories:
          type: array
          items:
            type: string
        alphabets:
          type: array
          items:
            type: string
        difficulty:
          type: ["string", "null"]
        totalQuestions:
          type: integer
          minimum: 5
          maximum: 100
        timeDurationMinutes:
          type: ["integer", "null"]
        questionMode:
          type: string
          enum: [random, alphabet_wise]
          default: random
    TestQuestion:
      type: object
      required: [id, position, questionText, options, vocabItemId]
      properties:
        id:
          type: integer
        position:
          type: integer
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        userAnswer:
          type: ["integer", "null"]
        vocabItemId:
          type: integer
    TestSummary:
      type: object
      required: [id, status, totalQuestions, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [in_progress, completed]
        totalQuestions:
          type: integer
        score:
          type: ["integer", "null"]
        timeDurationMinutes:
          type: ["integer", "null"]
        timeTakenSeconds:
          type: ["integer", "null"]
        config:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
    TestWithQuestions:
      type: object
      required: [id, status, totalQuestions, questions, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [in_progress, completed]
        totalQuestions:
          type: integer
        score:
          type: ["integer", "null"]
        timeDurationMinutes:
          type: ["integer", "null"]
        config:
          type: object
          additionalProperties: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/TestQuestion"
        createdAt:
          type: string
          format: date-time
    TestSubmission:
      type: object
      required: [answers, timeTakenSeconds]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuestionAnswer"
        timeTakenSeconds:
          type: integer
    QuestionAnswer:
      type: object
      required: [questionId, selectedOption]
      properties:
        questionId:
          type: integer
        selectedOption:
          type: ["integer", "null"]
    TestResultQuestion:
      type: object
      required: [id, position, questionText, options, correctAnswer, userAnswer, isCorrect, vocabItemId]
      properties:
        id:
          type: integer
        position:
          type: integer
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        userAnswer:
          type: ["integer", "null"]
        isCorrect:
          type: ["boolean", "null"]
        vocabItemId:
          type: integer
        vocabItem:
          $ref: "#/components/schemas/VocabularyItem"
    TestResult:
      type: object
      required: [id, score, totalQuestions, percentage, timeTakenSeconds, questions, createdAt, completedAt]
      properties:
        id:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        percentage:
          type: number
        timeTakenSeconds:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/TestResultQuestion"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    TestHistoryPage:
      type: object
      required: [items, total, page, limit, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TestSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    # ─── Progress Schemas ──────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      required: [wordsLearned, testsAttempted, currentStreak, averageScore, recentTests, dailyGoalProgress]
      properties:
        wordsLearned:
          type: integer
        testsAttempted:
          type: integer
        currentStreak:
          type: integer
        averageScore:
          type: number
        wordsToRevise:
          type: integer
        weakWordsCount:
          type: integer
        recentTests:
          type: array
          items:
            $ref: "#/components/schemas/TestSummary"
        dailyGoalProgress:
          type: integer
    LeaderboardEntry:
      type: object
      required: [rank, userId, name, wordsLearned, testsAttempted, averageScore]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        wordsLearned:
          type: integer
        testsAttempted:
          type: integer
        averageScore:
          type: number
...