# lib/api-zod/src/generated/types/testConfig.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SSC Vocabulary Master API
 * OpenAPI spec version: 0.1.0
 */
import type { TestConfigQuestionMode } from './testConfigQuestionMode';

export interface TestConfig {
  categories?: string[];
  alphabets?: string[];
  /** @nullable */
  difficulty?: string | null;
  /**
     * @minimum 5
     * @maximum 100
     */
  totalQuestions: number;
  /** @nullable */
  timeDurationMinutes?: number | null;
  questionMode?: TestConfigQuestionMode;
}