# lib/api-zod/src/generated/types/testHistoryPage.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SSC Vocabulary Master API
 * OpenAPI spec version: 0.1.0
 */
import type { TestSummary } from './testSummary';

export interface TestHistoryPage {
  items: TestSummary[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}