# lib/api-zod/src/generated/types/testResultQuestion.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SSC Vocabulary Master API
 * OpenAPI spec version: 0.1.0
 */
import type { VocabularyItem } from './vocabularyItem';

export interface TestResultQuestion {
  id: number;
  position: number;
  questionText: string;
  options: string[];
  correctAnswer: number;
  /** @nullable */
  userAnswer: number | null;
  /** @nullable */
  isCorrect: boolean | null;
  vocabItemId: number;
  vocabItem?: VocabularyItem;
}