# lib/api-zod/src/generated/types/testSubmission.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SSC Vocabulary Master API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionAnswer } from './questionAnswer';

export interface TestSubmission {
  answers: QuestionAnswer[];
  timeTakenSeconds: number;
}